Handle fetching integer and boolean keys when given a string key
Pretty much every interaction with getv will be a string path (e.g. /foo/bar), even when it contains numbers like /foo/123/bip. When unmarshalling yaml, however, the data structure stores their real types. When looking up in the map it will never find the string version of the keys. This change searches the map for the integer and boolean versions of the key to ensure we find integer keys.

We should probably rewrite the yaml maps to be map[string]any just like json, which would solve this problem from the source without the need for these shenanigans, but that's a larger effort.